1 Langchain/1.2 oLlama/SimpleAPP.py
import os 
from dotenv import load_dotenv
load_dotenv()

os.environ["LANGCHAIN_API_KEY"] = os.getenv('LANGCHAIN_API_KEY')
os.environ["LANGCHAIN_PROJECT"] = os.getenv("LANGCHAIN_PROJECT")
os.environ["LANGCHAIN_TRACING_V2"] = "true"

os.environ["USER_AGENT"] = (
    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) "
    "AppleWebKit/537.36 (KHTML, like Gecko) "
    "Chrome/113.0.0.0 Safari/537.36"
)

from langchain_ollama import OllamaLLM
import streamlit as st

from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import StrOutputParser

from langchain_core.prompts import SystemMessagePromptTemplate, HumanMessagePromptTemplate

prompt = ChatPromptTemplate.from_messages([
    SystemMessagePromptTemplate.from_template("You are a helpful assistant. Please respond to the question asked."),
    HumanMessagePromptTemplate.from_template("Question: {question}")
])
  

### streamlit framework
st.title("Langchain Demo With Gemma")
input_text = st.text_input("What question you have in mind?")

## Ollama Llama3
llm = OllamaLLM(model="gemma:2b")
output_parser = StrOutputParser()
chain = prompt | llm | output_parser
 
if input_text:
    st.write(chain.invoke({"question":input_text}))


### "1 Langchain/1.2 oLlama/SimpleAPP.py"